test: allow it to be both a test and main
